Use sed instead of ex in make_signature script
The make_signature script was using ex, which is part of vi editor and
seems to be less portable than sed, particularly when considering BSD
systems (e.g. FreeBSD). It lacks some of the GNU/Linux improvements,
causing failures of the script. We also do not check existence of the
command in configure, so it may be missing entirely.

Switching to sed fixes both those issues - it seems more portable, and
we already check it's availability in configure.

Patch contributed by John Schneider <john.a.schneider@gmail.com>,
additional comments and input by Jov <amutu@amutu.com>.