Make float8_tbl replicated to stabilize float8 test
As the table has just a single float8 column, XL automatically picks
ROUNDROBIN distribution. Commit 1d14325822 randomized selection of the
initial node, which with single-row inserts (used by the float8 tests)
effectively means random distribution, while before that all the rows
would be routed to the first node.

Some of the tests in float8 test suite seem to be sensitive to this, in
particular this overflow test:

    SELECT '' AS bad, f.f1 ^ '1e200' from FLOAT8_TBL f ORDER BY f1;
    ERROR:  value out of range: overflow

One of the rows (containing 1.2345678901234e-200) triggers an underflow,
so when the database hits it first, it will report this error instead:

    ERROR:  value out of range: underflow

The probability of hitting this is however fairly low (less than 10%),
as the executor has to touch the underflowing value first.