Fix plpgsql regression test
There were two broad categories of problems.

1. Errors due to lack of savepoint support
2. Errors and side effects due to lack of trigger support.

For 1, we reorganised the test case so that they can be run without savepoint.
For 2, we mostly accepted the regression changes. Apart from usual errors while
creating/dropping triggers, there were differences in query results because
of changes to preceding update/insert/delete statements. The behaviour of those
statements change because of lack of triggers.