Fix the pg_basebackup call when adding standby nodes
When adding standby nodes using pgxc_ctl, it's calling pg_basebackup
internally. But the "-x" option was removed in PostgreSQL 10, so the
call is failing.

A straightforward fix would be to use "-X fetch" which does exactly
what "-x" used to do. But I've decided to use "-X stream" instead,
as that does not rely on wal_keep_segments.

Reported by Tank.zhang"<6220104@qq.com>, along with "-X fetch" fix.