Stabilize combocid test by replicating the table
Commit 1d14325822 randomized the choice of a starting node with
ROUNDROBIN distribution, so that the data in combocid tests are not
all placed on the first node but distributed randomly (despite using
single-row INSERTS as before).

So to stabilize the test, make the table replicated. The table only
has a single column and the test updates is, so we can't use any
other stable distribution (e.g. BY HASH).

The expected results were obtained by running the combocid.sql on
PostgreSQL 10, so there might be some cmin differences.