Use local variable to format timestamp in GTM log line prefix
When formatting log line prefix in GTM, we can't use global variable,
because multiple threads may scribble over the same value. This is
why the timestamp was missing in some log lines - one thread did the
strftime(), but before it used the value another thread truncated the
string (which is the first step in formatting a log line).

So instead use a local (not shared by threads) variable, and pass it
to setup_formatted_log_time() explicitly.