Accept regression diffs post merge with 9.5.12
Two of the diffs are simply because of the addition of a Remote Subquery Scan
node on top of the plan. But two other diffs are actually error messages
resulting from a bug. But this bug is an old bug and there are already expected
output files with similar errors, hence accepting the diff for the time being.