Give a different treatment to brin test case.
Since function invokations are not automatically sent down to the remote node,
brin_summarize_range() doesn't report the correct result. Instead we now use
EXECUTE DIRECT mechanism to execute the function on the remote node and get the
result back. The expected output is adjusted acoordingly. Also move
xc_create_function test case at the beginning to ensure the function to query
remote node is available for this test case.