Accept regression diffs in collate test case
The differences are similar to XL 9.5 and not regressions. The root cause for
the difference is that in XL one CREATE TABLE AS SELECT runs without error,
whereas it throws an ERROR in vanilla Postgres. So the number of dependent
objects dropped at the end changes.

The way CREATE TABLE AS SELECT is implemented in XL, we split it into two steps
of CREATE TABLE and INSERT INTO. CREATE TABLE simply derives the type
information from the SELECT query, and hence unlike Postgres, it doesn't fail.